+++ CHANGELOG.md
@@ -4,10 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-## [NEXT] - - 
+## [NEXT] - -
+
+### Added
+- PubMed and RRID fields
 
 ### Changed
 - ORCID, ROR and PFAS icons are inlined.
+
+### Fixed
+- Multi instance problem with external authority fields.
 
 ## [1.5.0] - - 2025-05-30